refactor(backend): 🚚 rename client export
