+++ apps/backend/src/app.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from "bun:test";
 import { testClient } from "hono/testing";
 import app from "src/app";
-import type { ClientType } from "src/types";
+import type { ClientType } from "src/client";
 
 const { api } = testClient(app) as ClientType;
